config/apm: fix parsing DD_APM_FEATURES (#17630)
Support either "," or " " as separator when parsing the value of DD_APM_FEATURES. It fixes a regression introduced in #15904 which changed the separator from comma to space. This was a breaking change. From 7.44 to 7.46 using a space as separator was suggested ad a workaround, this PR ensures we don't break compatibility again. We now support either space or comma.