Double Agent replicate counts
We intend to automatically expand the number of replicates if a possible wobble
is detected. In order to approach the goal of doing this automatically we first
need to validate that consistently running this number of replicates is
1. feasible and 2. we draw more clear results by this doubling. There is work
required on our side to adjust the statistics, so in some sense this PR is a
change that will lead to a change that will lead to a change.

REF SMP-599
REF SMP-333

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>