Mark trace-agent Regression Detector experiments as erratic
Now that we have #17449 in place any experiment that wobbles without cause --
see #17672 -- needs looked into. We believe we have a bug in our rig that's
expressing in trace-agent experiments. To avoid dinging folks' PRs without cause
and not deny us follow-up data we set these experiments as 'erratic', meaning
they will still run but any detected "regression" will not be flagged.

We are actively pursuing this on our side. Much obliged to @hannahkm for the
report.

REF SMP-606

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>