Update lading version, trace-agent stable throttle
This commit updates the Agent CI's lading version to the latest release and
marks the throttle for trace-agent experiments to 'stable'. By default lading
uses a 'predictive' throttle, meaning we will examine the target's ingest rate
and push load to the capacity of the target. However, we do not participate in
back-off signals from the trace-agent yet -- see SMP-330 -- so the predictive
throttle is wide-open. We believe that SMP-606 -- see #17688 -- is due at least
in part to this mis-throttle.

REF SMP-330
REF SMP-606

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>