Fix race condition in the test
Occasionally, the test tcp connection would be closed while the test
body is still executing, causing a race condtion between assert on
listener.tailers and stopTailer.

Since the test assumes that the connection stays open until the end of
the test, make sure it stays open by adding an explicit close at the
end of the test body.