Prefix CI pulumi stack name by `ci-`
to avoid this error:
```
Your `clusterName` isn’t valid. It must be dot-separated tokens where a token start with a lowercase letter followed by lowercase letters, numbers, or hyphens, can only end with a with [a-z0-9] and has to be below 80 chars.
```
caused by the fact that cluster name is built from the pulumi stack name.