usm: sowatcher: Refactored initial scan mechanism
We read /proc/<pid>/smaps file to get the paths loaded into the process memory.
However, the smaps file is just an extension to /proc/<pid>/maps file, and all
the information relevant for our use case, can be obtained from maps file.

Furthermore, both files can have duplicated entries, and for us it is redundant
as we want to hook each file once per a matching .so file. So introducing a cache
to prevent duplications.

Another improvement, if a fast filtering for anonymous entries.