[CSPM] Handle system characteristics in XCCDF evaluator (#17966)
This change adds a function to OpenSCAP to print
system characteristics. This function is called
from the oscap-io tool when passing the "-syschar"
parameter.

The XCCDF evaluator has been modified to pass the
"-syschar" parameter when calling the oscap-io tool
and parse the system characteristics output.

The system characteristics are sent to the backend
as part of the data array.

This feature is disabled by default.