[Single-Machine-Performance] Update Regression Detector lading
This commit updates the lading version in the Regression Detector flow to
0.17.0. The major change the collection CPU utilization in lading. Previously we
collected tick information but absent precise measurement collection intervals
and core count it was not possible to say what the CPU utilization of a target
was. This proved to be important when examining the captured data from #17740 as
we could not see a distinct regression in tick data, despite there being one
present.

REF SMP-631

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>