Migrate `pkg/telemetry` to a Component (#17596)
This PR introduces a Telemetry component. The telemetry component is completely backwards compatible with existing telemetry counters which should make migration relatively painless. With this change, pkg/telemetry is now a thin proxy for comp/core/telemetry. This change also adds a new API to acquire an OTEL Meter. Meters will emit metrics to the same prometheus exporter as the rest of the agent telemetry.