fix: trim root folder suffix and trailing '/' when creating the zip files list
filepath.Join trim trailing '/' and as a result we could not query for directories
since they have a trailing '/' in their names.

We could have not use filepath.Join but instead let's remove trailing '/' everywhere
so that looking for 'path' or 'path/' is exactly the same thing

Also remove the root folder name from all filenames as it adds unnecessary clutter.

Signed-off-by: Nicolas Guerguadj <nico.guer@hotmail.com>