pkg/trace: resolve `agent status` bug on start up (#18123)
This PR fixes a bug in which running agent status returns an 
error when run shortly after starting the trace-agent. The bug 
was caused since the receiver was initialized as nil for the first 
10 seconds of the trace-agent starting up. The receiver's initial 
value is now an empty slice. This can help with debugging the 
agent and providing clarity to users.