[Serverless] AAS Universal Extension Support (#18023)
This PR adds support for the Azure App Services Universal Extension. Removing the `DD_RUNTIME` env variable in favor of just detecting the runtime based on the existence/values of other variables. Adding the logic to tag the version of the AAS extension here (currently being done in each tracer).