Optimize marshalling codepath in networks check
This commit brings in a version of the agent payload which uses [protoc-gen-gostreamer](https://github.com/leeavital/protoc-gen-gostreamer/).

This can avoid many intermediate allocations when
serializing network connections over the system-probe UDS. For
background, when the system-probe responds to a /connection RPC, the
following happens:

1. connections stored in state.go are pulled out as a
  network.Connections which under the hood is a big list of
  []network.Connection
2. the encoding package converts this []network.Connection to
  model.Connections, which under the hood is a []model.Connection with
  several buffers that hold compact representations of tags and DNS data
  as `[]byte`s
3. Then we call proto.Marshal on the model.Connection object
4. ...and write that into a response writer

In terms of allocations, step 2 creates a lot of short lived small
objects since each network.Connection needs to be converted to a
model.Connection. Alongside with a number of larger allocations for DNS
and tag sets. Before this PR, formatConnection incurred an allocation
per-connection each time an RPC happened.

Step 3 does one *big* allocation since the whole model.Connections object
needs to be allocated in a contiguous block of memory. Before this PR,
this happened in the Marshal method in protoSerializer.

The new code generated for protobuf marshalling allows different parts
of the network.Connections object to be written in separate calls. I.e.
instead of converting each network.Connection to model.Connection
object, each field in network.Connection can be written to the protobuf
separately without incurring a heap allocation.

This means a lot of encoding code needs to be re-written from
looking like this:

```
func FormatFoo(f *network.Foo) *model.Foo {
  return &model.Foo{
    x: int32(f.x),
  }
}
```

To

```
func FormatFoo(f *network.Foo, builder *model.FooBuilder) {
  builder.SetX(int32(f.x))
}
```

See FormatConnections for the most comprehensive example of how this is done.

[Benchmarks](https://ddstaging.datadoghq.com/notebook/5983918/perftest-fast-encode?range=1800000&start=1689038044502&live=false)
show a 36% memory improvement in the best case and 28% in the average
case in a load test. CPU impact appears to be negligible.