gitlab: rework deployment jobs grouping [AP-2154]
This moves from a v6/v7 split toward a container/packages split.
We already don't wait for previous jobs to complete when deploying
packages, so all jobs in this stage can use needs instead of
dependencies and start ASAP, regardless of the previous stages.
For the container jobs, we want to wait for the previous stages to
complete successfully.
Moving all containers deployment in the same stage allows for those jobs
to run in parallel