[Single Machine Performance] Update `smp` and `lading`
This commit is a peer to #18502. We update the `smp` binary to the most recent
version allowing the new nightly-only tool to be used in the future and also
update lading to the latest release. Of note, this lading is able to collect
process-specific CPU utilization data, elsewise contains small bug-fixes.

REF SMP-673

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>