Explore alternate methods to compute Workload Checks time
It turns out that the CI_COMMIT_TIME we introduced on nightly container builds
stopped functioning at some point between that introduction and the introduction
of the functional check. It always felt brittle and I guess it actually
was. This commit is an attempt too compute CI commit time in a manner that
doesn't rely on a job definition that is broadly shared but is local to the
workflow where that information is needed.

REF SMP-673

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>