Only build nightly SMP images for `main` branch
The Workload Checks analysis is built with the assumption that we will be
comparing successive main branches. We misunderstood how the nightly scheduled
functioned and believed that we would be running from simply main. This is not
true. To correct this we restrict builds of
single_machine_performance-nightly-amd64-a7 to main branch only.

REF SMP-673

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>