Avoid unnecessary allocations in `BuildAgentKey`
In the regression detector experiment `tcp_syslog_to_blackhole` we see that
`BuildAgentKey` accounts for ~20% of Agent allocations, see [this
notebook](https://app.datadoghq.com/notebook/6376068/agent-performance-investigation-gethostname?view=view-mode)
for reference to the profiles. Most of _those_ allocations are in `path.Join`
and as the Agent never has a '..' et al present in its keys and the prefix is
guaranteed non-empty we can simplify to just the use of a `strings.Builder`.

REF SMP-673

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>