Use a bytes buffer internally in the logs prefixer
The logs prefixer prepends a prefix to passed bytes per-destination. Our
[profiles](https://app.datadoghq.com/profiling/search?query=account%3Asingle-machine-performance%20client_team%3Aagent%20runtime%3Ago%20&my_code=disabled&profile_type=alloc-samples&viz=flame_graph&start=1693503706271&end=1693507306271&paused=false)
flag the `apply` function as allocation center. Considering that each
destination maintains an instantiated `prefixer` we can elide some of that
allocation pressure through the use of a buffer.

Note also we avoid converting the passed `prefix` from a string to bytes,
storing it as a string internally.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>