[omnibus] Share dpkg cache across multiple threads in OpenSCAP
This change modifies the dpkginfo-init patch to share the dpkg
cache across multiple threads.

The previous approach was to allocate an instance of the dpkg
cache per thread, but it didn't work as expected, due to
the leaky nature of the apt library.

The new approach consists of allocating pkgCacheFile once,
but calling ReadOnlyOpen every time dpkginfo_get_by_name is called.
The Close method is called in dpkginfo_probe_fini.

Consequently, the package list is always up to date, without
having to free pkgCacheFile.