Seek to reuse allocations in dogstatsd's intern mechanism
This commit makes two changes to the dogstatsd string intern mechanism. When
creating a new instance of stringInterner we now size the strings map up to its
maximum size. In addition when we must purge the strings map we now loop through
the map and delete each value, preferring to retain and reuse the allocation.

Profiles suggest purges are relatively rare -- or, at least, they don't show up
in the profiles -- and it is growing the map that is the time cost.

REF SMP-677

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>