[CSPM] Compliance checks: cross-container inputs resolving
Adds the capability to resolve inputs related to a given container ID.

When a Resolver in created with ContainerID as option, we try and detect the
associated processes that run on this container. We then resolve file contents
related to the mounted root path of this process.

We also link to the resulting check events some metadata about which container
and image was associated with the check.