Truncate log lines into a shared buffer
This commit is an attempt to reduce the amount of slice growth that's present in
our profiles of log ingest. When we trim lines we allocate storage space for the
trimming which, if we truncate enough, shows up non-trivially in CPU time. This
maintains a buffer to avoid that realloc and also attempts to avoid appends
whenever possible.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>