Expand the UDS/DogStatsD experiment set
This commit updates the `uds_dogstatsd_to_api` to conform to the behavior in the
`_nodist` variant excepting that we maintain the ratio of metrics in the
`app_server` workload check. We also introduce a `_nomultivalue` variant of the
baseline experiment that is intended to probe Agent's behavior when multiple
values are not packed into the same metric payload.

REF SMP-664

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>