comp/otelcol: add collector component and otel-agent binary (#18360)
This change migrates the pkg/otlp collector to a reusable component and adds a test binary to validate reusability. It is still an intermediate step because some related items are not yet migrated to components (e.g. pkg/status) which continue to require the global state, and others are migrated to components but can not be depended on (e.g. logs component) because of lazy initialisations required by the collector (the logs channel).