[CSPM] Terminate oscap-io processes after benchmark run
This change adds a new function FinishXCCDFBenchmark,
called after all the rules from an XCCDF benchmark has
been run.

The FinishXCCDFBenchmark function terminates the oscap-io
processes, so they won't stay in memory when not necessary.

New oscap-io processes will be started during the next run
of an XCCDF benchmark.

This change also decreases the verbosity of log message
related to the end of processes, since it's normal behavior
now.