[CSPM] Fix leak of oscap-io processes in XCCDF evaluator (#19690)
This change fixes leak of oscap-io processes in XCCDF evaluator.

When timing out after waiting for results, we used to delete
the oscap-io process from the oscapIOs map.

However, this behavior was racing with the p.Stop function,
which is called at the end of the p.Run function, after the
oscap-io process was terminated. If the EvaluateXCCDFRule function
was called before the call to p.Stop, the process information
of the old process could have been overridden by the new process,
before it was finally deleted in p.Stop. In this case, it would
lead to the leak of the new oscap-io process.

This changes modifies the p.Kill function to wait for the p.DoneCh
channel to be closed, so we can be sure that p.Stop is called
before EvaluateXCCDFRule is called for a new rule.

This change also removes the various call to "delete(oscapIOs, p.File)",
so process information are always deleted in the p.Stop function.

Finally, we have to be careful not to call p.Kill with the
oscapIOs map locked, because p.Stop still have to delete
process informations from the oscapIOs map.