usm: gotls: reuse file registry (#19701)
* usm: gotls: Remove reudndant wrapper to unregisterProcess

* usm: gotls: Remove redundant wrapper for process monitor

The scope of process monitor is only within the start method, there is no need to save the process monitor and the cleaner callbacks
to later clean them, but we can and should use the defer cleaner in the main goroutine

* usm: gotls: Remove reudndant fields from runningBinary struct

Removed the binID field, as we have a map that translates binID to a running binary.
Removed mtime field, as it is redundant check

* usm: gotls: Simplify constructor

* usm: gotls: Remove redundant conversion

* usm: Expose path identifier fields

* usm: gotls: Remove duplicated struct

We had binary id struct, which is similar to PathIdentifier struct, but instead of having a single u64 field, we
split it to major and minor (each 32 bits). The change was to remove the redundant struct and align the code
with path identifier, both kernel and user mode

* usm: gotls: Remove duplicated call for NewPathIdentifier

* usm: gotls: moving to file registry - changing add/remove-inspections functions

During the transition into file registry, we need to detach the hooking functions from being methods of the go tls program.
The first step is add/remove-inspection functions. Instead of being methods, we moved them into functions that get the relevant map
as an argument.

* usm: gotls: moving to file registry - changing add/detach-hooks functions

During the transition into file registry, we need to detach the hooking functions from being methods of the go tls program.
The first step is add/detach-hooks functions. Instead of being methods, we moved them into functions that get the relevant mgr
as an argument.

* usm: gotls: Migrate gotls to use file registry

This commit is doing the heavy lifting of converting old logic into file registry.
We don't need to maintain anything related to process registration, or block cache
as those are now handled by the file registry. Furthermore, added register and
unregister callback creator to fit into file registry requirements."

* usm: gotls: Change analysis metric into prometheus

* usm: gotls: Fixed metric name