Add check to dwarfutil type finding logic to include slices (#19758)
Go implements certain types as structs internally. For example, slices.
Slices have 3 fields, `array`, `len`, and `cap`. In the Go implementation
of Go DI that we are building, we rely on DWARF for type definitions and
on the dwarfutils library (and bininspect) for calculating offsets of
various parameters. This all means that in order for us to properly
calculate offsets we need this library to be aware of the multiple fields
of the internal slice type and their fields.

This adds a check in `FindStructFieldOffset` to check if a given struct
type name is actually the name of a slice type.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>