[CSPM] Add low priority scheduling of benchmarks.
Adding a new scheduling interval for benchmarks that has a bigger default
value and shall be used to schedule benchmarks for which we want to lower
the impact on the resource usage of the machine.

By introducing this new timings, I had to rework the jittering that space out
randomly the benchmarks runs. Specifically to handle the first run of our
benchmark which is always required to run in a relatively short time to handle
short-lived hosts.