[windows] Conditionally add apm tracing to agent (#19662)
* [windows] Conditionally add apm tracing to agent

Conditionally adds apm tracing to the main agent installer.
Whether component is configured is determined by the entry in
`release.json`.

When configured, adds the apm injection component via merge module,
and sets up beginning configuration.

Note:  This extends previous work to allow component to be
conditionally added.  As noted in the "comment" entry in release.json,
it is not intended to be included in release versions, and no
artifacts _will_ be included in release versions as long as the
`release.json` entries are not included.

add files in OG installer to make tests pass

Fill out apm-inject config template

Fix a small typo

add release-signed version

* update to latest package; fix merge error

* fix mismatched if/end