[windows] fix startup with empty api key
Previous PR introduced a dependency on the trace agent config
component.  The Windows agent expects to be able to start (with
errors in the log etc.) to be able to start even without a valid
api key.

With the new dependency on the trace config, make the trace component,
on windows, able to initialize so that the Windows agent can start
accordingly