[NPM-2998] clean up sources of potential flakiness
We recently saw TestConntrackExists6 flake with the following stack
trace

```
       [4.19.0-25-arm64] /go/src/github.com/DataDog/datadog-agent/pkg/network/netlink/conntrack_integration_test.go:176
       [4.19.0-25-arm64] Error:      	Received unexpected error:
       [4.19.0-25-arm64] dial tcp6 [fd00::2]:80: i/o timeout
       [4.19.0-25-arm64] Test:       	TestConntrackExists6
```

The likely explanation for this is that the TCP server started by this
test didn't come up by the time PingTCP was called.

This change updates the StartTCPServer and STartUDPServer test helpers
to wait until the test servers can accept a ping before returning. This
should make all tests relying on StartTCPServer/StartUDPServer less
flaky.