[SMP] Introduce overhead experiments
This commit introduces a series of new regression detector experiments derived
from the `uds_dogstatsd_to_api` that are intended to inspect overhead per metric
as a matter of throughput. We may which to remove distribution metrics
considering their throughput related issues although I believe this is
interesting to experiment with up front.

REF SMPTNG-23
REF SMPTNG-24

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>