[NPM-3010] avoid potential flakes in TestConntrackExists
We saw a spike in failures for TestConntrackExists in which
conntrack.Exists() returned false for UDP connections.

My guess is this is because the UDP connections didn't make it all the
way through the kernel before checking if the connection exists.

This PR aims to solve the problem by having PingUDP wait for a response
from the test UDP server to ensure the conntrack entry is initialized.