[SMP] Add more _nodist uds experiments (#20036)
This commit extracts experiments from #19882 and #19990, expanding the
throughput parameter across the range of concern. Note that at the lower end of
the throughput range the full contexts will not be explored, meaning that
throughput per second is the dominate metric here. Contexts per second we will
need to derive from captured target telemetry.

Intended to back ongoing experiments with the stringInterner.

REF SMPTNG-12

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>