[omnibus] Work around memory leak with usage of libapt in OpenSCAP
On some systems, we noticed memory leak with our usage
of the APT library.

This change modifies the dpkginfo probe to remove the
call to cgCache->Close in the dpkginfo_get_by_name
function. This works around the memory leak, which is
caused by the re-loading of the dpkg cache after
cgCache->Close has been called.

Functionally, this change is not correct, because
the result of dpkginfo_get_by_name will be based on
a cache of the dpkg database, which could potentially
be outdated.

However, since the oscap-io process is now terminated
each execution of an XCCDF benchmark, a new cache will
be loaded with during the next execution of the oscap-io
process.