[omnibus] Fix premature termination of probe workers in OpenSCAP (#19991) (#20095)
On some systems, we noticed that the oscap-io tool was leaking
a large amount of memory, because of a premature termination of
probe workers.

The dpkginfo probe calls the pkgCacheFile->GetPkgCache method
to build the APT cache. The list of the architectures the system
supports should be defined in the "APT::Architectures" configuration
option. If it's not the case, the APT library executes
"/usr/bin/dpkg --print-foreign-architectures" to retrieve the list
of supported architectures.

However, for some reason, the thread executing the dpkg tool leads
to a conflict with the probe worker threads, causing their premature
termination and leaking memory.

This change fixes this issue by recopying the "APT::Architecture"
configuration into "APT::Architectures" when the list of supported
architectures is empty.