Add support for filtering tags in k8s workloads.
Just add the anotation tags.datadog.com/disable: "comma,separated,tags,to,be,disabled"

Add support in taglist to remove tags.

The idea is to use this before calling the tags.Compute() method
The current implementation is done for k8s workloads, but it could
be expanded to other places.