Add experiments focused on dogstatsd's string interner
This commit replaces the _nodist variants of the `uds_dogstatsd_to_api`
experiment with a set of parameterized string interner experiments, varying both
total contexts and throughput along two scales. The goal here is to understand
how, at 8MiB/sec throughput, the variation of contexts impacts the
stringInterner hotspot and along throughput to achieve coverage on a small range
of contexts.

REF SMPTNG-12
REF #19852
REF #19882
REF #19990

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>