usm: protocols: Introduce a new method - IsBuildModeSupported.
The new method tells if the protocol is supported by a given protocol. This is preliminary change for migrating gotls into a protocol.
All protocols besides go-tls are supported by all build modes. Currently the method is not being called, as we need to do a further change
before calling this method