usm: protocols: Reduced code duplication
Introducing a new method 'runForProtocol' which gets a protocol list, a callback, a callback cleaner and a phase name, and
runs the given callback for every protocol. If we succeed, then we keep the protocol in a new list, and eventually returnning
that list to the caller. In case of a callback failure, we call the cleaner.