usm: protocols: Make ebpf program to be aware of programs not supported by build mode
In this commit we're introducing 2 helper functions - getProtocolsForBuildMode and configureManagerWithSupportedProtocols
The first is filtering out the enabled protocols but not supported by the build mode. So we have 2 lists -
1. Supported list - enabled protocols which support the current build mode
2. Not supported list - contains the disabled protocols, and the enabled protocols which are not supported
   by the current build mode.

The other method is adding for every supported protocols the maps, probes and tail calls in the program's list
and providing a cleanup method (which removes the maps, probes and tailcalls we've just added) in case of a failure.

Finally, we have to move the initialization of the instructions patcher to be inside the `init` method, as the final
list of probes is determind there.