usm: protocols: move protocols to ebpf manager (#20465)
* usm: protocols: Move getProtocolStats to ebpfProgram

* usm: protocols: Move protocols termination to ebpfProgram

* usm: protocols: Reduced code duplication

Introducing a new method 'runForProtocol' which gets a protocol list, a callback, a callback cleaner and a phase name, and
runs the given callback for every protocol. If we succeed, then we keep the protocol in a new list, and eventually returnning
that list to the caller. In case of a callback failure, we call the cleaner.

* usm: protocols: Moved runForProtocol to ebpfProgram, and used it in Close

Moved the method runForProtocol to ebpf_main.go file, and used it in Close method

* usm: protocols: Move protocol start into ebpfProgram

Migrated the calls for PreStart and PostStart to be inside ebpfProgram Start method.
Also, deleted enabledProtocol member from Monitor as it is not used anymore

* usm: protocols: Migrated protocols initiation into ebpfProgram

* Fixed CR comments