usm: http2: Bump max interesting frames limit to 30
Trying to increase the current limit ends with a verification error of hitting the 1m instruction
limit. Thus, we had to look for conditions to simplify and remove to cut the number of branches
the verifier has to go through.

We can remove the if-clause in `skip_preface` which checks the boundaries before the call
as in case of exceeding the packet limits, `bpf_skb_load_bytes` will fail, and the buffer will
be empty, or at least, won't match the HTTP2 magic we're comparing to.

Removing the condition allows us to bump the limit up to 22 max interesting frames, without having
any other compilation or verification issues. Trying to bump the limit to 23, returns a compilation
error caused by exceeding the stack limitation (as we're creating `http2_tail_call_state_t` on the
stack).

To fix that, we're introducing a per-cpu array `http2_frames_to_process` which acts as a heap.
With that change we're able to bump the limit to 30. We're now limited by the max number
of nested tail calls allowed. The nubmer is 32 (the actual limit is 33, but some JITs don't
have the update), and we already have `socket__protocol_dispatcher` and `socket__http2_filter`.