[USM] make protocols init aware of build mode (#20492)
* usm: protocols: Added Instance field to ProtocolSpec

The field represents a real instance created by the factory. This is use to associate the instance and the spec,
so we can better control creation and deletion in the ebpf program lifecycle. This simplifies the future changes
To move protocol managament from the monitor to the ebpf program.

* usm: protocols: Changed the way we save the build mode.

Preliminary change, instead of setting the build mode after successful initialization, we set it upfront
so we'll know within  method, what is the current buid mode we're trying to use.
In case of an error, we're reseting it using a defer function

* usm: protocols: Make ebpf program to be aware of programs not supported by build mode

In this commit we're introducing 2 helper functions - getProtocolsForBuildMode and configureManagerWithSupportedProtocols
The first is filtering out the enabled protocols but not supported by the build mode. So we have 2 lists -
1. Supported list - enabled protocols which support the current build mode
2. Not supported list - contains the disabled protocols, and the enabled protocols which are not supported
   by the current build mode.

The other method is adding for every supported protocols the maps, probes and tail calls in the program's list
and providing a cleanup method (which removes the maps, probes and tailcalls we've just added) in case of a failure.

Finally, we have to move the initialization of the instructions patcher to be inside the `init` method, as the final
list of probes is determind there.