[usm] parse multiple frames in a single tail call (#20496)
* usm: http2: Remove reudndant check

The method `process_headers` halts when we process all interesting frames, so if we have 0 interesting
frames, it will halt in the beginning without processing anything, thus the condition is reundant and
add additional branches to the verifier

* usm: http2: Remove reudndant checks

For every frame we're overriding the `data_off` field at the very beginning, so there is no meanning to advance the `data_off` fields
in those cases. Also the we should above only when we consume all interesting headers, instead when reaching to the
end of the packet

* usm: http2: Restructure of http2_tail_call_state_t

Moving the array to the beginning spares a couple of padding bytes.

* usm: http2: Fix indentation

* usm: http2: reduce stack allocation pressure

Preliminary change to reduce a stack pressure.

* usm: http2: Allow multi-frame parsing

After all preparations, we can now introduce a loop to parse frames (up to 5) in a single packet

* usm: http2: Increase max headers for parsing to 25