Add command to show top dogstatsd contexts
Different from dogstatsd-stats, this shows the state of the
aggregator, rather than the incoming traffic and thus more useful when
troubleshooting memory usage.

As this is mostly interesting when free memory is low and number of
contexts is high, write the data into a file rather than passing it in
the api response directly. The file can be removed from the system
under pressure and analyzed on a different host even if the pod
crashes.