Reduce allocation pressure in process_linux
This commit attempts to reduce allocation pressure in process_linux, following
on from #20502. That PR indicates that ParseStatusLine is an allocation
hotspot. We attempt to avoid that hotspot by not converting the key, value
arguments to parseStatusKV -- a child function -- into strings but pass them as
`[]byte`. This allows us to elide conversion of the key into a string in all
case and the value in a few. If we have a `ParseInt` routine in the code base
that works directly on byte slices this result can be improved.

I have also introduced a new benchmark function, patterned on the existing test
function.

Previous:

```
BenchmarkParseStatusLine-16     421380	  2657 ns/op    984 B/op    50 allocs/op
```

Current:
```
BenchmarkParseStatusLine-16     734079    1446 ns/op    296 B/op    6 allocs/op
````

REF SMPTNG-74

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>