[pkg/network] Attempt to fix direction for intra-host outgoing connections
When the system-probe starts up there is a race that occurs where port mappings are missing when they are bound after the `/proc/net` state is read but before the probes are registered.  When that happens, the connections will be marked as outgoing.  We see this very frequently with a lot of agent traffic as the agent container is starting at the same time as the system probe.

This commit attempts to mitigate that by checking if an outgoing connection is from a non-ephemeral port to an ephemeral port on an intra-host connection.