Make tests using 8.8.8.8 more reliable
All the tests which use 8.8.8.8 went through a function called
testDnsStats. Which would look for a single timeout, failure, or
success.

In the cases that it was looking for a success or failure (NXDOMAIN), it was
possible for it to flake in one of two ways:
* 8.8.8.8 would not respond due to packet loss or slow responses
* 8.8.8.8 would respond, but the response would not be seen in time by
  the snooper before the test called getConnections(tracer).

This updates the test in two ways to avoid these flakes:
1. it wraps the test in a loop which retries the request to DNS and
   asserts that the correct error code
2. It adds a test only method to dns_snooper called WaitDomain() wich
   waits for a domain to make its way to the DNS stat keeper. The tests
   will:
    * clear the tracer state
    * run the DNS request
    * call waitDomain()
    * call getConnections(tracer)
  Using this order means getConnections() will have a fully processed
  DNS request.