AMLII-176 add prometheus based core telemetry
This patch adds a new core check to send default telemetry from the
agent while using the existing telemetry component for
instrumentation.

This is designed to replace StatsTelemetry, thus only a subset of
metric types required by existing telemetry is supported.

As StatsTelemetry was unconditional, this check is also enabled by
default in all installations.