Pre-allocate `eventsBySourceType` to events length
This commit pre-allocates a temporary type that is used only temporarily when
marshalling events to JSON in a non-streaming fashion. A comment in the code
suggests that the function will be removed when v2 of the API is available, so
perhaps this function does need removed now.

In any event, profiles suggest it's an allocation/CPU center. This change only
marginally improves the micro-benchmarks, shaving 1/8 of the time off, from 8
seconds to 7, without adjusting allocations.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>