[CSPM] Add a system-probe module to perform compliance checks
This PR adds a new module to the system-probe binary in order to perform
compliance checks accross containers. Specifically to fetch containerized DB
applications configurations on which we can perform compliance checks.

The module exposes an endpoint that loads and parses the configuration files
of a few known databases applications: Cassandra, MondoDB and PostgreSQL. It
then returns the parsed data to the caller, security-agent in our case. The
configuration file is located either by parsing the command line arguments of
the associated DB process (identified by its PID) or looking at default known
paths in /etc.

Performance impact:

The overall overhead of this module should be minimal. Also it will not be
called by default in the release to have time to QA the impact.

1. adding dependency on pkg/compliance/dbconfig only adds ~63KB to the
   unstripped binary size
2. compliance module's endpoint should only be called one every three hours by
   our security-agent.
3. the actual workload of this module only consists of loading a few
   configuration files of typically ~10s of kilobytes and perform a basic
   parsing to structure the data in JSON. Guardrails are in place to make sure
   we do not read more that 1MB of data per config files.